+++ docs/CHANGELOG.md
@@ -3,4 +3,4 @@
 ----------
 
 * ***6/5/18 - Initial Commit***
- - *Added RVTools 3.10.1*
+ - *Added RVTools 3.10.1, VMware vCenter Converter Standalone 6.2.0.1*